[ADD][FEATURE] add multi terminal feature
FEATURES:
* multiple terminal buffers can be handled on the same window
* there are 3 public APIs available
 - open(id)
 - close()
 - toggle()

open() - takes terminal number as parameter. terminal number is an unique
id that represents the terminal buffer. when the terminal id is passed,
the particular terminal should be opened. if there are no terminals with
the ID, a new terminal buffer will be created. when the terminal window
is in opened state, the terminal buffer should be focused

close() closes the terminal window

toggle() open and close the terminal window based on the current status